Move tests from defaults_test.go to respective _types_test.go files
